🐛(common) fix pylti for wsgi servers behind a TLS termination proxy
WSGI servers format their headers by prefixing them with HTTP_, capitalizing
all letters and replacing "-" by "_". Pylti is checking the value of the
"X-Forwarded-Proto" header with the http syntax for this header and was
missing it when the http header is named "HTTP_X_FORWARDED_PROTO" as is the
case in Django for example.